Microservers-CarRentalSystem/src/auth_service/requirements.txt
aiofiles==22.1.0
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
peewee==3.15.4
priority==2.0.0
psycopg2==2.9.5
quart==0.18.3
requests==2.28.1
toml==0.10.2
urllib3==1.26.13
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0
PyJWT
Flask
retrying
quart-cors